Add files via upload
An extension to the Caesar cipher that was thought to be unbreakable is the Vigenère cipher. In 
this cipher, each letter is shifted by a different amount. This cipher uses a “key," which consists 
of a number for each letter in the message that specifies how much to shift each letter by. 
For example, if the message is "keyboard" and the key is 16458524, the message would be 
encrypted as "lkcgwfth". 
A B C D E F G H I J K L M N O P Q R S T U V W X Y Z
C D E F G H I J K L M N O P Q R S T U V W X Y Z A B
Your task is to code the Vigenère cipher. You should have one method to encrypt a 
message and one to decrypt a message. 
The method to encrypt a message should: 
• Accept a message as an argument. You may assume the message is all lowercase letters. 
• Create a key by creating an array that has the same length of the message and filling it 
with random values 
• Shift each letter in the message by the corresponding amount in the key. Spaces should 
be left unchanged. 
• Print the message and the key. 
The method to decrypt a message should: 
• Accept a message and a key as arguments. 
• Decrypt the text by shifting each letter in the message by the corresponding amount in 
the key. 
• Print the resulting message